+++ test/cpu_ops_test.cpp
@@ -1317,7 +1317,7 @@ TEST_CASE(quant_conv2d_test)
     p.compile(migraphx::cpu::target{});
     auto result = p.eval({});
 
-    std::vector<float> s = {10197,
+    std::vector<int32_t> s = {10197,
                             10548,
                             11601,
                             11952,
@@ -1334,7 +1334,7 @@ TEST_CASE(quant_conv2d_test)
                             81666,
                             82746};
 
-    std::vector<float> results_vector;
+    std::vector<int32_t> results_vector;
     result.visit([&](auto output) { results_vector.assign(output.begin(), output.end()); });
     EXPECT(migraphx::verify_range(results_vector, s));
 }
@@ -1357,14 +1357,14 @@ TEST_CASE(quant_conv2d_test_default_mode)
     p.compile(migraphx::cpu::target{});
     auto result = p.eval({});
 
-    std::vector<float> s = {
+    std::vector<int32_t> s = {
         10197, 10548, 6939,  3420,  11601, 11952, 7839,  3852,  7383,  7590,  4953,  2421,  3480,
         3570,  2316,  1125,  25506, 26586, 17874, 9009,  29826, 30906, 20718, 10413, 20505, 21198,
         14187, 7119,  10527, 10860, 7257,  3636,  27045, 27396, 17739, 8604,  28449, 28800, 18639,
         9036,  17319, 17526, 11289, 5445,  7800,  7890,  5052,  2421,  77346, 78426, 52002, 25857,
         81666, 82746, 54846, 27261, 53769, 54462, 36075, 17919, 26511, 26844, 17769, 8820};
 
-    std::vector<float> results_vector;
+    std::vector<int32_t> results_vector;
     result.visit([&](auto output) { results_vector.assign(output.begin(), output.end()); });
     EXPECT(migraphx::verify_range(results_vector, s));
 }
@@ -1387,7 +1387,7 @@ TEST_CASE(quant_conv2d_test_valid_mode)
     p.compile(migraphx::cpu::target{});
     auto result = p.eval({});
 
-    std::vector<float> s = {10197,
+    std::vector<int32_t> s = {10197,
                             10548,
                             11601,
                             11952,
@@ -1404,7 +1404,7 @@ TEST_CASE(quant_conv2d_test_valid_mode)
                             81666,
                             82746};
 
-    std::vector<float> results_vector;
+    std::vector<int32_t> results_vector;
     result.visit([&](auto output) { results_vector.assign(output.begin(), output.end()); });
     EXPECT(migraphx::verify_range(results_vector, s));
 }
@@ -1423,14 +1423,14 @@ TEST_CASE(quant_conv2d_padding_test)
     p.add_instruction(migraphx::op::quant_convolution{{{1, 1}}, {{1, 1}}}, al, cl);
     p.compile(migraphx::cpu::target{});
     auto result          = p.eval({});
-    std::vector<float> s = {
+    std::vector<int32_t> s = {
         4521,  6753,  7014,  4635,  6858,  10197, 10548, 6939,  7830,  11601, 11952, 7839,  5007,
         7383,  7590,  4953,  10515, 15987, 16734, 11277, 16821, 25506, 26586, 17874, 19737, 29826,
         30906, 20718, 13593, 20505, 21198, 14187, 13161, 19281, 19542, 12699, 18522, 27045, 27396,
         17739, 19494, 28449, 28800, 18639, 11919, 17319, 17526, 11289, 34707, 51843, 52590, 34893,
         51813, 77346, 78426, 52002, 54729, 81666, 82746, 54846, 36057, 53769, 54462, 36075};
 
-    std::vector<float> results_vector;
+    std::vector<int32_t> results_vector;
     result.visit([&](auto output) { results_vector.assign(output.begin(), output.end()); });
     EXPECT(migraphx::verify_range(results_vector, s));
 }
@@ -1450,7 +1450,7 @@ TEST_CASE(quant_conv2d_padding_stride_test)
     p.compile(migraphx::cpu::target{});
     auto result = p.eval({});
 
-    std::vector<float> s = {4521,
+    std::vector<int32_t> s = {4521,
                             7014,
                             7830,
                             11952,
@@ -1466,7 +1466,7 @@ TEST_CASE(quant_conv2d_padding_stride_test)
                             52590,
                             54729,
                             82746};
-    std::vector<float> results_vector;
+    std::vector<int32_t> results_vector;
     result.visit([&](auto output) { results_vector.assign(output.begin(), output.end()); });
     EXPECT(migraphx::verify_range(results_vector, s));
 }
